dicom/requirements.txt
pylibjpeg 
pylibjpeg-libjpeg
pydicom
pillow
numpy